[Refactor] layout 컴포넌트 폴더 구조 컨벤션 적용 - layout/ 하위에 common/sa/ta/to/tu 역할별 폴더 구조 적용 - AdminLayout, BaseSidebar → common/ 이동 - 각 역할별 Layout, Sidebar → 해당 역할 폴더로 이동 - import 경로 수정
