+++ src/components/layout/index.ts
@@ -1,15 +1,19 @@
-// Base components
-export { AdminLayout } from './AdminLayout';
-export { BaseSidebar } from './BaseSidebar';
+// Common components
+export { AdminLayout } from './common/AdminLayout';
+export { BaseSidebar } from './common/BaseSidebar';
 
-// Role-specific sidebars
-export { SuperAdminSidebar } from './SuperAdminSidebar';
-export { TenantAdminSidebar } from './TenantAdminSidebar';
-export { TenantOperatorSidebar } from './TenantOperatorSidebar';
-export { TenantUserSidebar } from './TenantUserSidebar';
+// SA (System Admin)
+export { SuperAdminLayout } from './sa/SuperAdminLayout';
+export { SuperAdminSidebar } from './sa/SuperAdminSidebar';
 
-// Role-specific layouts
-export { SuperAdminLayout } from './SuperAdminLayout';
-export { TenantAdminLayout } from './TenantAdminLayout';
-export { TenantOperatorLayout } from './TenantOperatorLayout';
-export { TenantUserLayout } from './TenantUserLayout';
+// TA (Tenant Admin)
+export { TenantAdminLayout } from './ta/TenantAdminLayout';
+export { TenantAdminSidebar } from './ta/TenantAdminSidebar';
+
+// TO (Tenant Operator)
+export { TenantOperatorLayout } from './to/TenantOperatorLayout';
+export { TenantOperatorSidebar } from './to/TenantOperatorSidebar';
+
+// TU (Tenant User)
+export { TenantUserLayout } from './tu/TenantUserLayout';
+export { TenantUserSidebar } from './tu/TenantUserSidebar';
