+++ src/components/layout/tu/TenantUserSidebar/TenantUserSidebar.tsx
@@ -1,4 +1,4 @@
-import { BaseSidebar } from '../BaseSidebar';
+import { BaseSidebar } from '../../common/BaseSidebar';
 import { tenantUserMenuData, roleLabels } from '@/config/sidebar-menus';
 
 interface TenantUserSidebarProps {
